% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj.p.R
\name{adj.p}
\alias{adj.p}
\title{Function to calculate the adj_p based on the number of estimated effective tests}
\usage{
adj.p(vars_df, select = 0.05)
}
\arguments{
\item{vars_df}{data.frame with sampes as rows and variables as columns}

\item{select}{threshold for selection}
}
\description{
Function to calculate the adj_p based on the number of estimated effective tests. Hidden to user
}
