% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestNN.R
\name{bestNN}
\alias{bestNN}
\title{Function to select the best nnet, but this could be in the fit.nnetwork function}
\usage{
bestNN(xdata, Y, size = 2, vars.drop.frac = 0.2, verbose = FALSE)
}
\arguments{
\item{xdata}{data set to be analyzed}

\item{Y}{var to analyze}

\item{size}{nnet param: number of units in the hidden layer. Can be zero if there are skip-layer units}

\item{vars.drop.frac}{fraction of variables removed at each iteration step, it selects the most important variables 1-fraction at each step}
}
\description{
nnet is obtained with all vars, so-called oob error
1 - cor(predict(Nnet)[complete.cases(predict(Nnet)),1], Y[complete.cases(cbind(Y, xdata[, Nnet$coefnames]))])^2)
an iterated process consists on creating a new nnet with the (1-vars.drop.frac)*vars more important obtained in the previous step
and generating a new nnet model. The process stops when abs(mean(oob error)+sd(oob error)) is >= than previous step or only two vars are left
}
