% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestRF.R
\name{bestRF}
\alias{bestRF}
\title{Function to select the best rf, but this could be in the fit.rforest function
adapted from varSelRF (varSelRF Ramon Diaz-Uriarte, only for classification)}
\usage{
bestRF(xdata, Y, ntree = 100, vars.drop.frac = 0.2, verbose = FALSE)
}
\arguments{
\item{xdata}{data set to be analyzed}

\item{Y}{var to analyze}

\item{ntree}{number of trees}

\item{vars.drop.frac}{fraction of variables removed at each iteration step, it selects the most important variables 1-fraction at each step}
}
\description{
Random forest is obtained with all vars, oob error is computed and the most important variables
an iterated process consists on creating a new rf with the (1-vars.drop.frac)*vars more important obtained in the previous step
and generating a new rf model. The process stops when abs(mean(oob error)+sd(oob error)) is >= than previous step or only two vars are left
}
