% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.imputation.R
\name{check.imputation}
\alias{check.imputation}
\title{Function to check whether data needs imputation: NO CAL}
\usage{
check.imputation(vars_df, perc_imp = 70, imp_method)
}
\arguments{
\item{vars_df}{data.frame to be checked for completeness (if it needs imputation)}

\item{perc}{number in (0,100) that is required for each variable to be complete}
}
\description{
Function to check whether data needs imputation: NO CAL
}
