% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.methods.R
\name{compare.methods}
\alias{compare.methods}
\title{filtering for an object of class nlAssoc}
\usage{
compare.methods(list.meth, filter = TRUE, param = "p", comp = "below",
  threshold = 0.05)
}
\arguments{
\item{list.meth}{a NAMED list of nlAssoc objects as obtained in fit.model(), usually applying different methods on the same data}

\item{filter}{if nlAssoc object has not been previously filtered it can be done here. Default is TRUE}

\item{param}{parameter of the table in results to filter with. Default is p}

\item{comp}{comparison param comp threshold, p below 0.05 default. Default is below}

\item{threshold}{numerical, the threshold to select associated variables related to the specified param. Default is 0.05}
}
\description{
filtering for an object of class nlAssoc
}
