% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.nlAssoc.R
\name{filter.nlAssoc}
\alias{filter.nlAssoc}
\title{filtering for an object of class nlAssoc}
\usage{
filter.nlAssoc(res, param = "p", comp = "below", threshold = 0.05,
  sort.results = TRUE)
}
\arguments{
\item{res}{an nlAssoc object as obtained in fit.model()}

\item{param}{parameter of the table in results to filter with. Default is p}

\item{comp}{comparison param comp threshold, p below 0.05 default. Default is below}

\item{threshold}{numerical, the threshold to select associated variables related to the specified param. Default is 0.05}
}
\description{
filtering for an object of class nlAssoc
}
