% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.mfp.R
\name{fit.mfp}
\alias{fit.mfp}
\title{Function to fit a multivariable fractional polynomial model to a data.frame or an expressionSet}
\usage{
fit.mfp(data_m = data_m, vars_df = vars_df, df = 4, select = 0.05,
  select_adj = NULL, cores = 1L, ...)
}
\arguments{
\item{data_m}{\code{ExpressionSet} or data.frame with samples as columns and observations as rows could be an ExpressionSet(modif!)}

\item{vars_df}{data.frame with sampes as rows and variables as columns}

\item{df}{degrees of freedom for mfp, default 4, 1 would be linear}

\item{select}{alpha: sets the FP selection level for all predictors. Values for individual predictors may be changed via the fp function in the formula.}

\item{select_adj}{select: sets the variable selection level for all predictors. Values for individual predictors may be changed via the fp function in the formula.}

\item{cores}{cores in case of parallelization (no windows)}

\item{imp_method}{imputation method for missing values in variables}
}
\value{
a file results with model, variables
}
\description{
Fits and mfp model for each row in data using mfp package
}
\examples{
to be built

}
