% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.model.R
\name{fit.model}
\alias{fit.model}
\title{Fits the specified multivariable model to a data.frame or an expressionSet}
\usage{
fit.model(data_m, vars_df, model = c("dsa", "ewas", "mfp", "gam", "gamboost",
  "rforest", "nnetwork"), imp_method = "fastpmm", perc_imp = 70,
  cores = 1L, ...)
}
\arguments{
\item{data_m}{\code{ExpressionSet} or data.frame with samples as columns and observations as rows could be an ExpressionSet(modif!). rownames!}

\item{vars_df}{data.frame with samples as rows and variables as columns}

\item{imp_method}{imputation method for missing values in variables}

\item{perc_imp}{}

\item{cores}{cores in case of parallelization (no windows)}

\item{df}{degrees of freedom for mfp, default 4, 1 would be linear}

\item{select}{alpha: sets the FP selection level for all predictors. Values for individual predictors may be changed via the fp function in the formula.}

\item{select_adj}{select: sets the variable selection level for all predictors. Values for individual predictors may be changed via the fp function in the formula.}
}
\value{
a file results with model, variables
}
\description{
Prepares data and fits specified model using specific call for each case
}
\examples{
to be built
}
