% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.nnetwork.R
\name{fit.nnetwork}
\alias{fit.nnetwork}
\title{Function to fit neural network to a data.frame or an expressionSet}
\usage{
fit.nnetwork(data = data_m, vars_df = vars_df, size = 2L, cores = 1L,
  verbose = FALSE, ...)
}
\arguments{
\item{data}{\code{ExpressionSet} or data.frame with samples as columns and observations as rows could be an ExpressionSet(modif!)}

\item{vars_df}{data.frame with sampes as rows and variables as columns}

\item{size}{nnet param: number of units in the hidden layer. Can be zero if there are skip-layer units}

\item{cores}{cores in case of parallelization (no windows)}

\item{verbose}{logical to verbose (comment) the steps of the function, default(FALSE)}
}
\description{
Neural networks are adjusted for each row in data using package nnet
}
