% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.partDSA.R
\name{fit.partDSA}
\alias{fit.partDSA}
\title{Function to fit partDSA with boosting to a data.frame or an expressionSet}
\usage{
fit.partDSA(data = data_m, vars_df = vars_df, df = 3, cores = 1L, ...)
}
\arguments{
\item{data}{\code{ExpressionSet} or data.frame with samples as columns and observations as rows could be an ExpressionSet(modif!)}

\item{vars_df}{data.frame with sampes as rows and variables as columns}

\item{df}{degrees of freedom to apply to model}

\item{cores}{cores in case of parallelization (no windows)}
}
\description{
Fits a partDSA model using boosting for each row in data using package partDSA
}
