% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.rforest.R
\name{fit.rforest}
\alias{fit.rforest}
\title{Function to fit rforest to a data.frame or an expressionSet}
\usage{
fit.rforest(data = data_m, vars_df = vars_df, seed = NULL,
  maxnodes = 15, ntree = 100, cores = 1L, ...)
}
\arguments{
\item{data}{\code{ExpressionSet} or data.frame with samples as columns and observations as rows could be an ExpressionSet(modif!)}

\item{vars_df}{data.frame with sampes as rows and variables as columns}

\item{cores}{cores in case of parallelization (no windows)}

\item{df}{degrees of freedom to apply to model}
}
\description{
Random forest models are fitted for each row in data using package randomForest
}
