% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignif.R
\name{getSignif}
\alias{getSignif}
\title{filtering for an object of class nlAssoc, to get significative results}
\usage{
getSignif(res, param = "adj.p", comp = "below", threshold = 0.05,
  sort.results = TRUE)
}
\arguments{
\item{res}{an nlAssoc object as obtained by nlOmicAssoc()}

\item{param}{parameter of the table to filter with (vars_n, aic, Cor2, p or adj.p). Default is "adj.p"}

\item{comp}{comparison param comp threshold. Default is "below"}

\item{threshold}{numerical, the threshold to select associated variables related to the specified param. Default is 0.05}
}
\description{
filtering for an object of class nlAssoc, to get significative results
}
