% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.nlAssoc.R
\name{plot.nlAssoc}
\alias{plot.nlAssoc}
\title{plot for an object of class nlAssoc}
\usage{
\method{plot}{nlAssoc}(res, top.probes = 10, pdf.name = NULL)
}
\arguments{
\item{res}{an nlAssoc object as obtained by nlOmicAssoc}

\item{top.probes}{numeric, number of probes to plot,}

\item{pdf.name}{name and path of pdf file to store plots. Default NULL}
}
\description{
plot for an object of class nlAssoc
}
