% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.single.mfp.R
\name{plot.single.mfp}
\alias{plot.single.mfp}
\title{plot for an object of class nlAssoc, obtatined with model mfp, for just a feature}
\usage{
\method{plot}{single.mfp}(res, probe, xlim, ylim, seed = 1234)
}
\arguments{
\item{res}{an nlAssoc object as obtained by nlOmicAssoc}

\item{xlim}{xlim for plot}

\item{ylim}{ylim for plot}

\item{seed}{seed for internal random processes}

\item{probes}{character, id of the probe to plot,}
}
\description{
plot for an object of class nlAssoc, obtatined with model mfp, for just a feature
}
