% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.nlAssoc.R
\name{print.nlAssoc}
\alias{print.nlAssoc}
\title{Print for an object of class nlAssoc showing only first registries}
\usage{
\method{print}{nlAssoc}(res)
}
\arguments{
\item{res}{an nlAssoc object as obtained in fit.model()}
}
\description{
Print for an object of class nlAssoc showing only first registries
}
