% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.vars.nlAssoc.R
\name{score.vars.nlAssoc}
\alias{score.vars.nlAssoc}
\title{scoring the variables of selected_vars slot in nlAssoc object}
\usage{
score.vars.nlAssoc(res)
}
\arguments{
\item{res}{an nlAssoc object as obtained in fit.model()}
}
\description{
scoring the variables of selected_vars slot in nlAssoc object
}
