% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreVars.R
\name{scoreVars}
\alias{scoreVars}
\title{scoring the variables of selected_vars slot in an nlAssoc object}
\usage{
scoreVars(res)
}
\arguments{
\item{res}{an nlAssoc object as obtained by nlOmicAssoc()}
}
\description{
scoring the variables of selected_vars slot in an nlAssoc object
}
