% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.nlAssoc.R
\name{summary.nlAssoc}
\alias{summary.nlAssoc}
\title{Summary for an object of class nlAssoc}
\usage{
\method{summary}{nlAssoc}(res)
}
\arguments{
\item{res}{an nlAssoc object as obtained in fit.model()}
}
\description{
Summary for an object of class nlAssoc
}
