% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.gam.R
\name{fit.gam}
\alias{fit.gam}
\title{Function to fit gam to a data.frame or an expressionSet}
\usage{
fit.gam(data = data_m, vars_df = vars_df, df = 4L, cores = 1L,
  verbose = FALSE, ...)
}
\arguments{
\item{data}{\code{ExpressionSet} or data.frame with samples as columns and observations as rows could be an ExpressionSet(modif!)}

\item{vars_df}{data.frame with sampes as rows and variables as columns}

\item{df}{degrees of freedom to apply to model}

\item{cores}{cores in case of parallelization (no windows)}

\item{verbose}{logical to verbose (comment) the steps of the function, default(FALSE)}
}
\description{
GAM models are fitted for each row in data using lm and backfitting using bruto function in mda package.
}
