% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.mfp.R
\name{fit.mfp}
\alias{fit.mfp}
\title{Function to fit a multivariable fractional polynomial model to a data.frame or an expressionSet}
\usage{
fit.mfp(data = data_m, vars_df = vars_df, df = 4L, alpha = 0.05,
  select = 0.05, cores = 1L, verbose = FALSE, ...)
}
\arguments{
\item{vars_df}{data.frame with sampes as rows and variables as columns}

\item{df}{degrees of freedom for mfp, default 4, 1 would be linear}

\item{alpha}{sets the FP selection level for all predictors. Values for individual predictors may be changed via the fp function in the formula.}

\item{select}{select: sets the variable selection level for all predictors. Values for individual predictors may be changed via the fp function in the formula.}

\item{cores}{cores in case of parallelization (no windows)}

\item{verbose}{logical to verbose (comment) the steps of the function, default(FALSE)}

\item{data_m}{\code{ExpressionSet} or data.frame with samples as columns and observations as rows could be an ExpressionSet(modif!)}
}
\description{
Fits and mfp model for each row in data using mfp package
}
\examples{
to be built

}
