% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.model.R
\name{fit.model}
\alias{fit.model}
\title{Fits the specified multivariable model to a data.frame or an expressionSet}
\usage{
fit.model(set, vars_df = NULL, model = c("dsa", "ewas", "mfp", "gam",
  "gamboost", "rforest", "nnetwork"), imp_method = "pmm", perc_imp = 70,
  cores = 1L, verbose = FALSE, ...)
}
\arguments{
\item{set}{numeric matrix, or \code{ExpressionSet} or \code{SummarizedExperiment} with samples as columns and observations as rows could be an ExpressionSet(modif!). rownames!}

\item{vars_df}{data.frame with samples as rows and variables as columns}

\item{imp_method}{imputation method for missing values in variables}

\item{perc_imp}{numerical in [0,100], the maximum % of a variable allowed to have Nas}

\item{cores}{cores in case of parallelization (no windows)}

\item{verbose}{logical to verbose (comment) the steps of the function, default(FALSE)}

\item{df}{degrees of freedom for mfp, default 4, 1 would be linear}

\item{select}{alpha: sets the FP selection level for all predictors. Values for individual predictors may be changed via the fp function in the formula.}

\item{select_adj}{select: sets the variable selection level for all predictors. Values for individual predictors may be changed via the fp function in the formula.}
}
\value{
nlAssoc

a file results with model, variables
}
\description{
Prepares data and fits specified model using specific call for each case
Returns an object on nlAssoc class
}
\examples{
to be built
}
