% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.rforest.R
\name{fit.rforest}
\alias{fit.rforest}
\title{Function to fit rforest to a data.frame or an expressionSet}
\usage{
fit.rforest(data = data_m, vars_df = vars_df, seed = NULL,
  maxnodes = 15L, ntree = 2000, cores = 1L, verbose = FALSE, ...)
}
\arguments{
\item{data}{\code{ExpressionSet} or data.frame with samples as columns and observations as rows could be an ExpressionSet(modif!)}

\item{vars_df}{data.frame with sampes as rows and variables as columns}

\item{seed}{set a see, for reproducibility}

\item{maxnodes}{maximum number of terminal nodes trees in the forest can have}

\item{ntree}{number of trees to grow}

\item{cores}{cores in case of parallelization (no windows)}

\item{verbose}{logical to verbose (comment) the steps of the function, default(FALSE)}

\item{df}{degrees of freedom to apply to model}
}
\description{
Random forest models are fitted for each row in data using package randomForest
}
